requirements.txt
openai>=1.50.0
langchain>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.12.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
